fix no thaumcraft crash
